Refactored unit tests.
